Pull out V1 namespace
