+++ app/controllers/api/v1x0/graphql_controller.rb
@@ -1,6 +1,18 @@
+require "insights/api/common/graphql"
+
 module Api
   module V1x0
-    class GraphqlController < Api::V1::GraphqlController
+    class GraphqlController < ApplicationController
+      def query
+        graphql_api_schema = ::Insights::API::Common::GraphQL::Generator.init_schema(request, overlay)
+        variables = ::Insights::API::Common::GraphQL.ensure_hash(params[:variables])
+        result = graphql_api_schema.execute(
+          params[:query],
+          :variables => variables
+        )
+        render :json => result
+      end
+
       def overlay
         {
           "^.*$" => {
