+++ app/controllers/api/v1x0/mixins/rbac_mixin.rb
@@ -1,5 +1,5 @@
 module Api
-  module V1
+  module V1x0
     module Mixins
       module RBACMixin
         include ApprovalPermissions
