+++ app/controllers/api/v1x0/stageaction/_footer.html.erb
