+++ app/controllers/api/v1x0/stageaction/result.html.erb
